Installer works!
Silent installer totally works. Success!

+++ Install-Office-2016-osx.sh
@@ -1,14 +1,19 @@
 #!/bin/bash -x
 
-##### This script will be used to Install Microsoft Office 2016.
+###### This is a simple script which will do a silent install of Office 2016.
 
-sudo -S mkdir /MS\ Office
-cd /MS\ Office/
+read -p "Do you want to install Microsoft Office 2016? y/n " choice
 
-# The link below doesn't seem to be working. 
-# Using this link seems to create a master that just opens itself infinitely when double-clicked: https://github.com/nuclearsalt/install-office-2016-osx/archive/master.zip
-# sudo -S curl -O https://codeload.github.com/nuclearsalt/install-office-2016-osx/zip/master
-open master.zip
-cd /MS\ Office/install-office-2016-osx-master/
-sudo -S curl -O https://officecdn-microsoft-com.akamaized.net/pr/C1297A47-86C4-4C1F-97FA-950631F94777/OfficeMac/Microsoft_Office_2016_16.15.18070902_Installer.pkg
-sudo -S /usr/sbin/installer -pkg '/MS\ Office/install-office-2016-osx-master/Microsoft_Office_2016_16.15.18070902_Installer.pkg' -target '/Volumes/Macintosh HD/' -applyChoiceChangesXML /MS\ Office/install-office-2016-osx-master/choices.plist
+	if [[ $choice = "Y" || $choice = "y" ]] ; then 		
+		sudo -S mkdir /MS\ Office
+		sudo -S cp choices.plist /MS\ Office/
+		cd /MS\ Office/
+		sudo -S curl -O https://officecdn-microsoft-com.akamaized.net/pr/C1297A47-86C4-4C1F-97FA-950631F94777/OfficeMac/Microsoft_Office_2016_16.15.18070902_Installer.pkg
+		echo "Downloading Office Installer" 
+		sudo -S /usr/sbin/installer -pkg '/MS Office/Microsoft_Office_2016_16.15.18070902_Installer.pkg' -target '/Volumes/Macintosh HD/' -applyChoiceChangesXML /MS\ Office/choices.plist
+		sudo -S mv /MS\ Office ~/.Trash
+		echo "Done!"
+	else
+		echo "Office will not be installed right now."
+		exit 0
+	fi
